training_pipeline/requirements.txt
mlflow>=1.0
scikit-learn==1.2.2
pandas==1.3.5
category-encoders==2.6.0
pycaret==3.0.0
threadpoolctl==3.1.0
pyspark==3.4.1
fugue==0.8.5
aniso8601==9.0.1
attrs==22.1.0
cfgv==3.3.1
click==8.1.3
coverage==6.5.0
distlib==0.3.6
exceptiongroup==1.0.0
filelock==3.8.0
greenlet==2.0.0
identify==2.5.8
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
marshmallow==3.18.0
mysqlclient==2.1.1
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
pytz==2022.6
PyYAML==6.0
SQLAlchemy==1.4.42
toml==0.10.2
tomli==2.0.1
virtualenv==20.16.6
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.10.0
mysql-connector-python
pendulum==2.1.2
apache-airflow
apache-airflow-providers-docker